+++ Resources/config/api.xml
@@ -5,14 +5,11 @@
     xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <services>
-        <service id="api.resource" class="Dunglas\ApiBundle\Api\Resource" public="false" abstract="true" />
+        <service id="api.resource" class="Dunglas\ApiBundle\Api\Resource" public="false" abstract="true" lazy="true" />
 
-        <service id="api.resource_collection" class="Dunglas\ApiBundle\Api\ResourceCollection" />
+        <service id="api.resource_collection" class="Dunglas\ApiBundle\Api\ResourceCollection" lazy="true" />
 
-        <service id="api.data_provider" class="Dunglas\ApiBundle\Model\DataProviderChain">
-            <argument type="service" id="api.resource_collection" />
-            <argument type="service" id="api.router" />
-        </service>
+        <service id="api.data_provider" class="Dunglas\ApiBundle\Model\DataProviderChain" lazy="true" />
 
         <service id="api.operation_factory" class="Dunglas\ApiBundle\Api\Operation\OperationFactory" public="false" />
 
@@ -25,7 +22,12 @@
 
         <service id="api.router" class="Dunglas\ApiBundle\Routing\Router" public="false">
             <argument type="service" id="router" />
+        </service>
+
+        <service id="api.iri_converter" class="Dunglas\ApiBundle\Api\IriConverter" public="false">
             <argument type="service" id="api.resource_collection" />
+            <argument type="service" id="api.data_provider" />
+            <argument type="service" id="api.router" />
             <argument type="service" id="property_accessor" />
         </service>
     </services>
