+++ Resources/config/doctrine_orm.xml
@@ -23,7 +23,7 @@
         </service>
 
         <service id="api.doctrine.orm.filter" class="Dunglas\ApiBundle\Doctrine\Orm\Filter" public="false" abstract="true">
-            <argument type="service" id="api.data_provider" />
+            <argument type="service" id="api.iri_converter" />
             <argument type="service" id="property_accessor" />
         </service>
     </services>
