+++ Routing/Router.php
@@ -11,8 +11,6 @@
 
 namespace Dunglas\ApiBundle\Routing;
 
-use Dunglas\ApiBundle\Api\ResourceCollectionInterface;
-use Dunglas\ApiBundle\Api\ResourceInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
 use Symfony\Component\Routing\RequestContext;
@@ -26,31 +24,13 @@ use Symfony\Component\Routing\RouterInterface;
 class Router implements RouterInterface
 {
     /**
-     * @var \SplObjectStorage
-     */
-    private $routeCache;
-    /**
      * @var RouterInterface
      */
     private $router;
-    /**
-     * @var ResourceCollectionInterface
-     */
-    private $resourceCollection;
-    /**
-     * @var PropertyAccessorInterface
-     */
-    private $propertyAccessor;
 
-    public function __construct(
-        RouterInterface $router,
-        ResourceCollectionInterface $resourceCollection,
-        PropertyAccessorInterface $propertyAccessor
-    ) {
+    public function __construct(RouterInterface $router)
+    {
         $this->router = $router;
-        $this->resourceCollection = $resourceCollection;
-        $this->propertyAccessor = $propertyAccessor;
-        $this->routeCache = new \SplObjectStorage();
     }
 
     /**
@@ -102,17 +82,6 @@ class Router implements RouterInterface
      */
     public function generate($name, $parameters = [], $referenceType = self::ABSOLUTE_PATH)
     {
-        if (is_object($name)) {
-            if ($name instanceof ResourceInterface) {
-                $name = $this->getCollectionRouteName($name);
-            } else {
-                if ($resource = $this->resourceCollection->getResourceForEntity($name)) {
-                    $parameters['id'] = $this->propertyAccessor->getValue($name, 'id');
-                    $name = $this->getItemRouteName($resource);
-                }
-            }
-        }
-
         $baseContext = $this->router->getContext();
 
         try {
@@ -128,72 +97,6 @@ class Router implements RouterInterface
             return $this->router->generate($name, $parameters, $referenceType);
         } finally {
             $this->router->setContext($baseContext);
-        }
-    }
-
-    /**
-     * Gets the collection route name for a resource.
-     *
-     * @param ResourceInterface $resource
-     *
-     * @return string
-     */
-    private function getCollectionRouteName(ResourceInterface $resource)
-    {
-        $this->initRouteCache($resource);
-
-        if (isset($this->routeCache[$resource]['collectionRouteName'])) {
-            return $this->routeCache[$resource]['collectionRouteName'];
-        }
-
-        $operations = $resource->getCollectionOperations();
-        foreach ($operations as $operation) {
-            if (in_array('GET', $operation->getRoute()->getMethods())) {
-                $data = $this->routeCache[$resource];
-                $data['collectionRouteName'] = $operation->getRouteName();
-                $this->routeCache[$resource] = $data;
-
-                return $data['collectionRouteName'];
-            }
-        }
-    }
-
-    /**
-     * Gets the item route name for a resource.
-     *
-     * @param ResourceInterface $resource
-     *
-     * @return string
-     */
-    private function getItemRouteName(ResourceInterface $resource)
-    {
-        $this->initRouteCache($resource);
-
-        if (isset($this->routeCache[$resource]['itemRouteName'])) {
-            return $this->routeCache[$resource]['itemRouteName'];
-        }
-
-        $operations = $resource->getitemOperations();
-        foreach ($operations as $operation) {
-            if (in_array('GET', $operation->getRoute()->getMethods())) {
-                $data = $this->routeCache[$resource];
-                $data['itemRouteName'] = $operation->getRouteName();
-                $this->routeCache[$resource] = $data;
-
-                return $data['itemRouteName'];
-            }
-        }
-    }
-
-    /**
-     * Initializes the route cache structure for the given resource.
-     *
-     * @param ResourceInterface $resource
-     */
-    private function initRouteCache(ResourceInterface $resource)
-    {
-        if (!$this->routeCache->contains($resource)) {
-            $this->routeCache[$resource] = [];
         }
     }
 }
